Fix problems in text source data conversion and fix tests

+++ v2/googlecloud-to-neo4j/src/main/java/com/google/cloud/teleport/v2/neo4j/model/enums/PropertyType.java
@@ -21,7 +21,7 @@ public enum PropertyType {
   Integer,
   Long,
   Float,
-  BigDecimal,
+  Double,
   Boolean,
   Date,
   LocalDateTime,
